+++ tests/conftest.py
@@ -124,3 +124,27 @@ def tmp_package_modi_2(root, tmp_empty):
     yield tmp_empty
 
     os.chdir(old)
+
+
+@pytest.fixture
+def tmp_package_changelog(root, tmp_empty):
+    old = Path.cwd()
+    path_to_templates = root / "tests" / "assets" / "package_name"
+    shutil.copytree(str(path_to_templates), "copy")
+    os.chdir("copy")
+
+    subprocess.run(["git", "init"])
+    subprocess.check_call(["git", "config", "commit.gpgsign", "false"])
+    subprocess.check_call(["git", "config", "user.email", "ci@ploomberio"])
+    subprocess.check_call(["git", "config", "user.name", "Ploomber"])
+    subprocess.run(["git", "checkout", "-b", "main"])
+    with open("CHANGELOG.md", "a") as f:
+        f.write("\n")
+    subprocess.run(["git", "add", "--all"])
+    subprocess.run(["git", "commit", "-m", "init-commit-message"])
+
+    subprocess.run(["git", "checkout", "-b", "test_modified_changelog"])
+
+    yield tmp_empty
+
+    os.chdir(old)