+++ tests/test_modified.py
@@ -1,5 +1,7 @@
 import pytest
-from pkgmt import fail_if_modified, fail_if_not_modified
+import subprocess
+
+from pkgmt import fail_if_modified, fail_if_not_modified, fail_if_invalid_changelog
 from pathlib import Path
 
 
@@ -81,3 +83,110 @@ def test_check_modified_in_2(tmp_package_modi_2, base_branch, include_path, retu
         fail_if_not_modified.check_modified(base_branch, include_path, debug=True)
         == returncode
     )
+
+
+# @pytest.mark.parametrize("contents", ["""
+# # CHANGELOG
+#
+# ## 0.1dev
+#
+# * [Fix] Fixes #1
+# * [Feature] Added feature 1
+# """
+# ])
+def test_check_modified_changelog(tmp_package_changelog):
+    Path("CHANGELOG.md").write_text(
+        """
+# CHANGELOG
+
+## 0.1dev
+
+* [Fix] Fixes #1
+* [Feature] Added feature 1
+"""
+    )
+    subprocess.run(["git", "add", "CHANGELOG.md"])
+    subprocess.run(["git", "commit", "-m", "changelog_modified"])
+    assert (
+        fail_if_invalid_changelog.check_modified("main", ["CHANGELOG.md"], debug=True)
+        == 0
+    )
+
+
+@pytest.mark.parametrize(
+    "contents, message",
+    [
+        (
+            """
+# CHANGELOG
+
+## 0.1dev
+
+* [Fix] Fixes #1
+
+
+""",
+            "CHANGELOG.md has not been modified with respect to 'main'",
+        ),
+        (
+            """
+# CHANGELOG
+
+## 0.1dev
+
+* [Fix] Fixes #1
+
+""",
+            "CHANGELOG.md has not been modified with respect to 'main'",
+        ),
+        (
+            """
+# CHANGELOG
+
+## 0.0.1dev
+
+* [Fix] Fixes #1
+
+""",
+            "Latest section in CHANGELOG.md not up-to-date. "
+            "Latest section in main: 0.1dev",
+        ),
+        (
+            """
+# CHANGELOG
+
+## 0.1dev
+
+* [Fix] Fixes #1
+
+## 0.0.1
+
+* [Feature] Some feature
+
+""",
+            "Entry '* [Feature] Some feature' should be added to section 0.1dev",
+        ),
+        (
+            """
+# CHANGELOG
+
+## 0.1dev
+
+* [Fix] Fixes #2
+
+""",
+            "These entries have been removed: * [Fix] Fixes #1. "
+            "Please revert the changes.",
+        ),
+    ],
+    ids=["blanks", "no-change", "outdated-header", "incorrect-section", "removal"],
+)
+def test_check_modified_changelog_error(
+    tmp_package_changelog, contents, message, capsys
+):
+    Path("CHANGELOG.md").write_text(contents)
+    subprocess.run(["git", "add", "CHANGELOG.md"])
+    subprocess.run(["git", "commit", "-m", "changelog_modified"])
+    assert fail_if_invalid_changelog.check_modified("main", debug=True) == 1
+    out = capsys.readouterr().out
+    assert message in out