+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.7.5dev
 * [Feature] Added `--version` option to `pkgmt setup` to specify Python version
+* [Feature] Better checks for new entries added to CHANGELOG.md
 
 ## 0.7.4 (2023-09-08)